+++ src/components/NavBar/NavBar.jsx
@@ -1,7 +1,12 @@
 import styles from './NavBar.module.css';
 import Auth from '../Auth/Auth';
+import { useContext } from 'react';
+import { UserContext } from '../../context/user.context';
 
-function NavBar({ loggedInUser, users, setUsers }) {
+function NavBar() {
+	const { users } = useContext(UserContext);
+	const loggedInUser = users.find((user) => user.isLogined === true);
+	const loggedInUserName = loggedInUser ? loggedInUser.userName : '';
 	return (
 		<div className={styles['navBar']}>
 			<div className={styles['logo']}>
@@ -15,7 +20,7 @@ function NavBar({ loggedInUser, users, setUsers }) {
 					Мои фильмы
 					<div className={styles['menu__counter']}>X</div>
 				</div>
-				<Auth loggedInUser={loggedInUser} users={users} setUsers={setUsers} />
+				<Auth loggedInUser={loggedInUserName} />
 			</div>
 		</div>
 	);
