+++ src/context/user.context.jsx
@@ -1,17 +1,23 @@
-import { createContext } from 'react';
-import { useLocalStorage } from '../hooks/use-localstorage.hook';
+import { createContext, useEffect, useState } from 'react';
 
 export const UserContext = createContext();
 
 export const UserProvider = ({ children }) => {
-	const [users, setUsers] = useLocalStorage('data', []);
+	const [users, setUsers] = useState(() => {
+		const storedData = localStorage.getItem('data');
+		return storedData ? JSON.parse(storedData) : [];
+	});
+
+	useEffect(() => {
+		localStorage.setItem('data', JSON.stringify(users));
+	}, [users]);
 
 	const handleLogout = () => {
 		setUsers(users.map((user) => ({ ...user, isLogined: false })));
 	};
 
 	return (
-		<UserContext.Provider value={{ handleLogout }}>
+		<UserContext.Provider value={{ users, setUsers, handleLogout }}>
 			{children}
 		</UserContext.Provider>
 	);
